Replace app.vue with index page
